+++ L38TRN_HFT_2021221.Logic/SongLogic.cs
@@ -11,14 +11,34 @@ namespace L38TRN_HFT_2021221.Logic
     public class SongLogic : ISongLogic
     {
         ISongRepository songRepo;
-        public void Update(int id, string newSongName)
+
+        public SongLogic(ISongRepository songRepo)
         {
-            songRepo.UpdateSongName(id, newSongName);
+            this.songRepo = songRepo;
+        }
+
+        public void Update(int id, Song song)
+        {
+            if (songRepo.GetOne(id) != null && song != null)
+            {
+                songRepo.Update(id, song);
+            }
+            else
+            {
+                throw new InvalidOperationException("ERROR");
+            }
         }
 
         public Song Read(int id)
         {
-            return songRepo.GetOne(id);
+            if (songRepo.GetOne(id) != null)
+            {
+                return songRepo.GetOne(id);
+            }
+            else
+            {
+                throw new InvalidOperationException("ERROR");
+            }
         }
 
         public List<Song> GetAll()
@@ -28,17 +48,45 @@ namespace L38TRN_HFT_2021221.Logic
 
         public void Create(Song song)
         {
-            songRepo.Create(song);
+            if (song != null)
+            {
+                songRepo.Create(song);
+            }
+            else
+            {
+                throw new InvalidOperationException("ERROR");
+            }  
         }
 
         public void Delete(int id)
         {
-            songRepo.DeleteSong(id);
+            if (songRepo.GetOne(id) != null)
+            {
+                songRepo.DeleteSong(id);
+            }
+            else
+            {
+                throw new InvalidOperationException("ERROR");
+            }
         }
 
         public IEnumerable<Song> ReadAll()
         {
             return songRepo.GetAll();
+        }
+
+        public IEnumerable<Song> MostListenedSongs()
+        {
+            return from x in songRepo.GetAll()
+                   where x.NumberOfListens > 150000
+                   select x;
+        }
+
+        public IEnumerable<Song> LeastListenedSongs()
+        {
+            return from x in songRepo.GetAll()
+                   where x.NumberOfListens < 1000
+                   select x;
         }
     }
 }
